+++ backend/Features/ShoppingLists/IShoppingListService.cs
@@ -12,5 +12,5 @@ internal interface IShoppingListService
   
   Task<UnitResult<Error>> RemoveSublist(int shoppingSublistId);
   
-  Task<UnitResult<Error>> UpdateSublistCount(int shoppingSublistId, int count);
+  Task<UnitResult<Error>> UpdateSublistCount(int shoppingSublistId, decimal count);
 }
